requirements.txt
opencv-python
torch
torchvision
numpy
matplotlib
pandas
opencv-python
pillow